requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.2
python-multipart>=0.0.24
langchain>=1.2.15
langchain-core>=1.2.26
langchain-community>=0.4.1
langchain-openai>=1.1.12
langchain-text-splitters>=1.1.1
langchain-classic>=1.0.3
langchain-neo4j>=0.9.0
pydantic>=2.12.5
litellm>=1.80.0
openai>=2.30.0
neo4j>=6.1.0
tiktoken>=0.12.0
requests>=2.33.0
wikipedia>=1.4.0
python-dotenv>=1.2.2
SQLAlchemy>=2.0.49
psycopg2-binary>=2.9.6